Refactored types
